Buildings and floating bodies done
